requirements.txt
# Django
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
drf-spectacular>=0.27.0

# Database
psycopg2-binary>=2.9.9

# Agent Framework
crewai>=0.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.6

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.2

# Static/Dynamic Analysis
requests>=2.31.0
pypdf>=4.0.0

# Browser Testing
playwright>=1.40.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
